+++ internal/domain/usecases/user/update_user_patient_id/interfaces.go
@@ -6,5 +6,5 @@ import (
 )
 
 type WriteRepo interface {
-	UpdateUserPatientId(ctx context.Context, user entity.User, patientId int64) (err error)
+	UpdateUserPatientId(ctx context.Context, user entity.User, patientId int) (err error)
 }
